Move everything into MacroParserImpl
