+++ src/jdk.incubator.jextract/share/classes/jdk/internal/jextract/impl/MacroParserImpl.java
@@ -26,6 +26,9 @@
 
 package jdk.internal.jextract.impl;
 
+import jdk.incubator.jextract.Declaration;
+import jdk.incubator.jextract.JextractTask;
+import jdk.incubator.jextract.Position;
 import jdk.incubator.jextract.Type;
 import jdk.internal.clang.Cursor;
 import jdk.internal.clang.CursorKind;
@@ -38,20 +41,24 @@ import jdk.internal.clang.TranslationUnit;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-class MacroParserImpl {
+class MacroParserImpl implements JextractTask.ConstantParser {
 
     private Reparser reparser;
-    TypeMaker typeMaker = new TypeMaker(null);
+    TreeMaker treeMaker = new TreeMaker();
 
-    public MacroParserImpl(TranslationUnit tu, Collection<String> args) {
+    public MacroParserImpl(TreeMaker treeMaker, TranslationUnit tu, Collection<String> args) {
         try {
             this.reparser = new ClangReparser(tu, args);
+            this.treeMaker = treeMaker;
         } catch (IOException | Index.ParsingFailedException ex) {
             this.reparser = Reparser.DUMMY;
         }
@@ -74,6 +81,97 @@ class MacroParserImpl {
         return Optional.empty();
     }
 
+    final Map<String, UnparsedMacro> macrosToReparse = new LinkedHashMap<>();
+
+    @Override
+    public Optional<Declaration.Constant> parseConstant(Position pos, String name, String[] tokens) {
+        if (!(pos instanceof TreeMaker.CursorPosition)) {
+            return Optional.empty();
+        } else {
+            Cursor cursor = ((TreeMaker.CursorPosition)pos).cursor();
+            if (cursor.isMacroFunctionLike()) {
+                return Optional.empty();
+            } else {
+                Optional<MacroParserImpl.Macro> macro = eval(name, tokens);
+                if (macro.isEmpty()) {
+                    macrosToReparse.put(name, new UnparsedMacro(name, tokens, cursor));
+                }
+                return macro.map(m -> treeMaker.createMacro(cursor, m));
+            }
+        }
+    }
+
+    class UnparsedMacro {
+        final String name;
+        final String[] tokens;
+        final Cursor cursor;
+
+        public UnparsedMacro(String name, String[] tokens, Cursor cursor) {
+            this.name = name;
+            this.tokens = tokens;
+            this.cursor = cursor;
+        }
+
+        String constantDecl(boolean forcePtr) {
+            //we use __auto_type, so that clang will also do type inference for us
+            return (forcePtr) ?
+                    "__auto_type jextract$macro$ptr$" + name + " = (uintptr_t)" + name + ";" :
+                    "__auto_type jextract$macro$" + name + " = " + name + ";";
+        }
+    }
+
+    String macroDecl(Collection<UnparsedMacro> macros, boolean forcePtr) {
+        StringBuilder buf = new StringBuilder();
+        if (forcePtr) {
+            buf.append("#include <stdint.h>\n");
+        }
+        for (UnparsedMacro m : macros) {
+            buf.append(m.constantDecl(forcePtr));
+            buf.append("\n");
+        }
+        return buf.toString();
+    }
+
+    public List<Declaration.Constant> reparseConstants() {
+        //slow path
+        List<Declaration.Constant> macros = new ArrayList<>();
+        //step one, parse constant as is
+        int last = -1;
+        while (macrosToReparse.size() > 0 && macrosToReparse.size() != last) {
+            Map<String, Type> pendingTypes = new LinkedHashMap<>();
+            Map<String, UnparsedMacro> pendingMacros = new LinkedHashMap<>();
+            last = macrosToReparse.size();
+            List<MacroParserImpl.MacroResult> results = reparse(macroDecl(macrosToReparse.values(), false));
+            for (MacroParserImpl.MacroResult result : results) {
+                UnparsedMacro unparsedMacro = macrosToReparse.get(normalizeName(result.name));
+                if (result.success()) {
+                    macros.add(treeMaker.createMacro(unparsedMacro.cursor, (MacroParserImpl.Macro) result));
+                    macrosToReparse.remove(normalizeName(result.name));
+                } else if (result.type != null) {
+                    pendingMacros.put(normalizeName(result.name), unparsedMacro);
+                    pendingTypes.put(normalizeName(result.name), result.type);
+                } else {
+                    // this is not a recoverable failure
+                    macrosToReparse.remove(normalizeName(result.name));
+                }
+            }
+            if (!pendingMacros.isEmpty()) {
+                //step two, attempt parsing pointer constant, by forcing it to uintptr_t
+                results = reparse(macroDecl(pendingMacros.values(), true));
+                for (MacroParserImpl.MacroResult result : results) {
+                    UnparsedMacro unparsedMacro = pendingMacros.get(normalizeName(result.name));
+                    if (result.success()) {
+                        result = result.asType(pendingTypes.get(normalizeName(result.name)));
+                        macros.add(treeMaker.createMacro(unparsedMacro.cursor, (MacroParserImpl.Macro) result));
+                    }
+                    // never look at this again
+                    macrosToReparse.remove(normalizeName(result.name));
+                }
+            }
+        }
+        return macros;
+    }
+
     List<MacroResult> reparse(String snippet) {
         try {
             return reparser.reparse(snippet)
@@ -82,7 +180,7 @@ class MacroParserImpl {
                     .map(c -> compute(c))
                     .collect(Collectors.toList());
         } finally {
-            typeMaker.resolveTypeReferences();
+            treeMaker.typeMaker.resolveTypeReferences();
         }
     }
 
@@ -100,19 +198,19 @@ class MacroParserImpl {
             switch (result.getKind()) {
                 case Integral: {
                     long value = result.getAsInt();
-                    return Macro.longMacro(decl.spelling(), typeMaker.makeType(decl.type()), value);
+                    return Macro.longMacro(decl.spelling(), treeMaker.typeMaker.makeType(decl.type()), value);
                 }
                 case FloatingPoint: {
                     double value = result.getAsFloat();
-                    return Macro.doubleMacro(decl.spelling(), typeMaker.makeType(decl.type()), value);
+                    return Macro.doubleMacro(decl.spelling(), treeMaker.typeMaker.makeType(decl.type()), value);
                 }
                 case StrLiteral: {
                     String value = result.getAsString();
-                    return Macro.stringMacro(decl.spelling(), typeMaker.makeType(decl.type()), value);
+                    return Macro.stringMacro(decl.spelling(), treeMaker.typeMaker.makeType(decl.type()), value);
                 }
                 default:
                     return new Failure(decl.spelling(), decl.type().equals(decl.type().canonicalType()) ?
-                            null : typeMaker.makeType(decl.type()));
+                            null : treeMaker.typeMaker.makeType(decl.type()));
             }
         }
     }
@@ -235,11 +333,7 @@ class MacroParserImpl {
         }
     }
 
-    static class BadMacroException extends RuntimeException {
-        static final long serialVersionUID = 1L;
-
-        public BadMacroException(Diagnostic diag) {
-            super(diag.toString());
-        }
+    static String normalizeName(String name) {
+        return name.substring(name.lastIndexOf('$') + 1);
     }
 }
