+++ src/jdk.incubator.jextract/share/classes/jdk/internal/jextract/impl/Parser.java
@@ -68,7 +68,7 @@ class Parser {
             true, args.toArray(new String[0]));
 
         JextractTask.ConstantParser constantParser = this.constantParser != null ?
-                this.constantParser : new DefaultConstantParser(new MacroParserImpl(tu, args));
+                this.constantParser : new MacroParserImpl(treeMaker, tu, args);
 
         List<Declaration> decls = new ArrayList<>();
         Cursor tuCursor = tu.getCursor();
@@ -107,8 +107,8 @@ class Parser {
                 }
             });
 
-        if (constantParser instanceof DefaultConstantParser) {
-            decls.addAll(((DefaultConstantParser)constantParser).reparseConstants());
+        if (constantParser instanceof MacroParserImpl) {
+            decls.addAll(((MacroParserImpl)constantParser).reparseConstants());
         }
         Declaration.Scoped rv = treeMaker.createHeader(tuCursor, decls);
         treeMaker.freeze();
@@ -118,108 +118,5 @@ class Parser {
 
     private boolean isMacro(Cursor c) {
         return c.isPreprocessing() && c.kind() == CursorKind.MacroDefinition;
-    }
-
-    class DefaultConstantParser implements JextractTask.ConstantParser {
-
-        final MacroParserImpl macroParser;
-        final Map<String, UnparsedMacro> macrosToReparse = new LinkedHashMap<>();
-
-        public DefaultConstantParser(MacroParserImpl macroParser) {
-            this.macroParser = macroParser;
-        }
-
-        @Override
-        public Optional<Declaration.Constant> parseConstant(Position pos, String name, String[] tokens) {
-            if (!(pos instanceof TreeMaker.CursorPosition)) {
-                return Optional.empty();
-            } else {
-                Cursor cursor = ((TreeMaker.CursorPosition)pos).cursor();
-                if (cursor.isMacroFunctionLike()) {
-                    return Optional.empty();
-                } else {
-                    Optional<MacroParserImpl.Macro> macro = macroParser.eval(name, tokens);
-                    if (macro.isEmpty()) {
-                        macrosToReparse.put(name, new UnparsedMacro(name, tokens, cursor));
-                    }
-                    return macro.map(m -> treeMaker.createMacro(cursor, m));
-                }
-            }
-        }
-
-        class UnparsedMacro {
-            final String name;
-            final String[] tokens;
-            final Cursor cursor;
-
-            public UnparsedMacro(String name, String[] tokens, Cursor cursor) {
-                this.name = name;
-                this.tokens = tokens;
-                this.cursor = cursor;
-            }
-
-            String constantDecl(boolean forcePtr) {
-                //we use __auto_type, so that clang will also do type inference for us
-                return (forcePtr) ?
-                        "__auto_type jextract$macro$ptr$" + name + " = (uintptr_t)" + name + ";" :
-                        "__auto_type jextract$macro$" + name + " = " + name + ";";
-            }
-        }
-
-        String macroDecl(Collection<UnparsedMacro> macros, boolean forcePtr) {
-            StringBuilder buf = new StringBuilder();
-            if (forcePtr) {
-                buf.append("#include <stdint.h>\n");
-            }
-            for (UnparsedMacro m : macros) {
-                buf.append(m.constantDecl(forcePtr));
-                buf.append("\n");
-            }
-            return buf.toString();
-        }
-
-        public List<Declaration.Constant> reparseConstants() {
-            //slow path
-            List<Declaration.Constant> macros = new ArrayList<>();
-            //step one, parse constant as is
-            int last = -1;
-            while (macrosToReparse.size() > 0 && macrosToReparse.size() != last) {
-                Map<String, Type> pendingTypes = new LinkedHashMap<>();
-                Map<String, UnparsedMacro> pendingMacros = new LinkedHashMap<>();
-                last = macrosToReparse.size();
-                List<MacroParserImpl.MacroResult> results = macroParser.reparse(macroDecl(macrosToReparse.values(), false));
-                for (MacroParserImpl.MacroResult result : results) {
-                    UnparsedMacro unparsedMacro = macrosToReparse.get(normalizeName(result.name));
-                    if (result.success()) {
-                        macros.add(treeMaker.createMacro(unparsedMacro.cursor, (MacroParserImpl.Macro) result));
-                        macrosToReparse.remove(normalizeName(result.name));
-                    } else if (result.type != null) {
-                        pendingMacros.put(normalizeName(result.name), unparsedMacro);
-                        pendingTypes.put(normalizeName(result.name), result.type);
-                    } else {
-                        // this is not a recoverable failure
-                        macrosToReparse.remove(normalizeName(result.name));
-                    }
-                }
-                if (!pendingMacros.isEmpty()) {
-                    //step two, attempt parsing pointer constant, by forcing it to uintptr_t
-                    results = macroParser.reparse(macroDecl(pendingMacros.values(), true));
-                    for (MacroParserImpl.MacroResult result : results) {
-                        UnparsedMacro unparsedMacro = pendingMacros.get(normalizeName(result.name));
-                        if (result.success()) {
-                            result = result.asType(pendingTypes.get(normalizeName(result.name)));
-                            macros.add(treeMaker.createMacro(unparsedMacro.cursor, (MacroParserImpl.Macro) result));
-                        }
-                        // never look at this again
-                        macrosToReparse.remove(normalizeName(result.name));
-                    }
-                }
-            }
-            return macros;
-        }
-    }
-
-    static String normalizeName(String name) {
-        return name.substring(name.lastIndexOf('$') + 1);
     }
 }
